Add JCS reference vectors
